feat: add TaEtiqueta component and integrate into routing
